java: fix conversion of ints and longs in PartialRowPB
A copy-paste error caused any ints/longs to get truncated to
shorts when serializing as PartialRowPB. This caused the
ITBLL cluster to end up writing almost all of its data into
just two tablets (the first and last) since the splits
ranged from -32000 to +32000 instead of -2^65 to +2^65.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7664
Tested-by: Todd Lipcon <todd@cloudera.com>

+++ java/kudu-client/src/main/java/org/kududb/client/PartialRow.java
@@ -496,12 +496,12 @@ public class PartialRow {
           break;
         }
         case INT32: {
-          short val = Bytes.getShort(this.rowAlloc, this.schema.getColumnOffset(idx));
+          int val = Bytes.getInt(this.rowAlloc, this.schema.getColumnOffset(idx));
           builder.addColumnsBuilder().setIndex(idx).setInt32Val(val);
           break;
         }
         case INT64: {
-          short val = Bytes.getShort(this.rowAlloc, this.schema.getColumnOffset(idx));
+          long val = Bytes.getLong(this.rowAlloc, this.schema.getColumnOffset(idx));
           builder.addColumnsBuilder().setIndex(idx).setInt64Val(val);
           break;
         }
